Moved to seperate package (potential lib classes)
